Fixing compilers warnings for print format and unused variables, uncommenting sampling functions

+++ src/networks/model-net/model-net.c
@@ -326,7 +326,7 @@ static model_net_event_return model_net_event_impl_base(
     tw_lpid dest_mn_lp = model_net_find_local_device_mctx(net_id, recv_map_ctx,
             final_dest_lp);
 
-    if (src_mn_lp == dest_mn_lp && message_size < codes_node_eager_limit)
+    if (src_mn_lp == dest_mn_lp && message_size < (uint64_t)codes_node_eager_limit)
         return model_net_noop_event(final_dest_lp, is_pull, offset, message_size,
                 remote_event_size, remote_event, self_event_size, self_event,
                 sender);
